% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hat_a.R
\name{a_hat}
\alias{a_hat}
\title{Non-parametric estimate of the acceleration factor for $BC_a$ confidence intervals}
\usage{
a_hat(bootdata)
}
\arguments{
\item{bootdata}{A numeric vector or matrix of bootstrap/ jackknife samples}
}
\value{
later
}
\description{
Non-parametric estimate of the acceleration factor for $BC_a$ confidence intervals
}
