% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apffc_gen.R
\name{Adap_gen}
\alias{Adap_gen}
\title{Generating an adapted progressive first-failure dataset from a given dataset}
\usage{
Adap_gen(m, R, Time, data)
}
\arguments{
\item{m}{number of desired failure}

\item{R}{Prefixed censoring scheme}

\item{Time}{Prefixed adaptation time}

\item{data}{a matrix, vector or a data frame. If the input is a vector, a matrix is created by taking the smallest factor of the length of the vector as columns.}
}
\value{
A list
}
\description{
Generating an adapted progressive first-failure dataset from a given dataset
}
