% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goftest.R
\name{AdapGoF}
\alias{AdapGoF}
\title{Approximate Goodness-of-fit test for progressive first-failure censored data}
\usage{
AdapGoF(data, CDF, k = 1, norm_approx = T, ...)
}
\arguments{
\item{data}{A 2 by m matrix, containing failure times in first row and the progressive first-failure censoring scheme in second row}

\item{CDF}{Cumulative distribution function}

\item{k}{number of units per group}

\item{norm_approx}{logical. Whether automatically calculate approximated goodness of fit test statistics}

\item{...}{further arguments to be passed to CDF. Typically this is a vector of parameter estimates}
}
\value{
later
}
\description{
Approximate Goodness-of-fit test for progressive first-failure censored data
}
