% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChenD.R
\name{Chen}
\alias{Chen}
\alias{dchen}
\alias{pchen}
\alias{qchen}
\alias{rchen}
\title{The Chen Distribution}
\usage{
dchen(x, shape, rate = 1, log = FALSE)

pchen(q, shape, rate = 1, lower.tail = TRUE, log.p = FALSE)

qchen(p, shape, rate = 1, lower.tail = TRUE, log.p = FALSE)

rchen(n, shape, rate = 1)
}
\arguments{
\item{x, q}{Vector of quantiles.}

\item{shape}{shape parameter.}

\item{rate}{rate parameter}

\item{log, log.p}{logical}

\item{lower.tail}{logical; if \code{TRUE}}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\description{
The Chen Distribution
}
