% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRD_CDF.R
\name{dGR}
\alias{dGR}
\alias{pGR}
\alias{qGR}
\alias{rGR}
\title{The Generalized Rayleigh Distribution}
\usage{
dGR(x, shape, scale = 1, rate = 1/scale, log = FALSE)

pGR(q, shape, scale = 1, rate = 1/scale, lower.tail = TRUE, log.p = FALSE)

qGR(p, shape, scale = 1, rate = 1/scale, lower.tail = TRUE)

rGR(n, shape, scale = 1, rate = 1/scale)
}
\arguments{
\item{x, q}{Vector of quantiles.}

\item{shape}{shape parameter.}

\item{scale}{scale parameter.}

\item{rate}{an alternative way to specify the scale.}

\item{log}{logical}

\item{lower.tail}{logical; if \code{TRUE}}

\item{log.p}{logical}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\description{
The Generalized Rayleigh Distribution
}
