% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_Lindley.R
\name{gLindley}
\alias{gLindley}
\alias{dgLindley}
\alias{pgLindley}
\alias{qgLindley}
\alias{rgLindley}
\title{The Generalized Lindley Distribution}
\usage{
dgLindley(x, shape = 1, rate, log = FALSE)

pgLindley(q, shape = 1, rate, lower.tail = TRUE, log.p = FALSE)

qgLindley(p, shape = 1, rate, lower.tail = TRUE, log.p = FALSE)

rgLindley(n, shape = 1, rate)
}
\arguments{
\item{x, q}{Vector of quantiles.}

\item{shape}{shape parameter.}

\item{rate}{rate parameter}

\item{log, log.p}{logical}

\item{lower.tail}{logical; if \code{TRUE}}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
later
}
\description{
The Generalized Lindley Distribution
}
