% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRD.R
\name{GR}
\alias{GR}
\alias{dgenrayl}
\alias{pgenrayl}
\alias{qgenrayl}
\alias{rgenrayl}
\title{The Generalized Rayleigh Distribution}
\usage{
dgenrayl(x, shape, rate = 1, scale = 1/rate, log = FALSE)

pgenrayl(q, shape, rate = 1, scale = 1/rate, lower.tail = TRUE, log.p = FALSE)

qgenrayl(p, shape, rate = 1, scale = 1/rate, lower.tail = TRUE, log.p = FALSE)

rgenrayl(n, shape, rate = 1, scale = 1/rate)
}
\arguments{
\item{x, q}{Vector of quantiles.}

\item{shape}{shape parameter.}

\item{rate}{an alternative way to specify the scale.}

\item{scale}{scale parameter.}

\item{log, log.p}{logical}

\item{lower.tail}{logical; if \code{TRUE}}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\description{
The Generalized Rayleigh Distribution
}
