% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRD_CDF.R
\name{pgr}
\alias{pgr}
\title{The Generalized Rayleigh Distribution}
\usage{
pgr(q, shape, scale = 1, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{q}{Vector of quantiles}

\item{shape}{shape parameter}

\item{scale}{scale parameter}

\item{lower.tail}{logical}

\item{log.p}{logical}
}
\description{
The Generalized Rayleigh Distribution
}
