% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFFCQuantile.R
\name{qpffc2}
\alias{qpffc2}
\title{Quantile Function of Progressive First-failure Censored Order Statistics}
\usage{
qpffc2(p = seq(0, 1, 0.25), n, m, k = 1, R, order, QF, ..., precBits = 1024)
}
\arguments{
\item{p}{probabilities}

\item{n}{number of experimental groups}

\item{m}{desired number of failures}

\item{k}{group sizes}

\item{R}{progressive censoring scheme}

\item{order}{order of the progressive first-failure censored order statistic}

\item{QF}{quantile function of the distribution}

\item{...}{additional arguments to be passed to \code{QF}}

\item{precBits}{same as \code{Rmpfr}}
}
\value{
Later
}
\description{
Quantile Function of Progressive First-failure Censored Order Statistics
}
