% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apffc2.R
\name{rapffc2}
\alias{rapffc2}
\title{Title}
\usage{
rapffc2(
  n,
  m = NULL,
  k = 1,
  Time,
  R = NULL,
  prob = NULL,
  CDF,
  QF,
  ...,
  set_seed = NULL
)
}
\arguments{
\item{n}{Total number of groups placed on test at time zero}

\item{m}{Number of required failures}

\item{k}{Number of items in a group}

\item{Time}{Prefixed time point}

\item{R}{Prefixed censoring scheme}

\item{prob}{Probability of removal}

\item{CDF}{Cumulative Distribution function of the univariate distribution}

\item{QF}{Quantile function of the univariate distribution}

\item{...}{Further arguments to be passed to \code{CDF} and \code{QF}}

\item{set_seed}{Random number generation seed}
}
\value{
Laetr
}
\description{
Title
}
