% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAPFFC.R
\name{rgapffc2}
\alias{rgapffc2}
\title{Simulating a generalized adaptive progressively first-failure type-II censored sample from a univariate distribution}
\usage{
rgapffc2(
  n,
  m = NULL,
  k = 1,
  Time1 = Inf,
  Time2 = Inf,
  R = NULL,
  non_zero = T,
  prob = NULL,
  CDF,
  QF,
  ...,
  set_seed = NULL
)
}
\arguments{
\item{n}{Total number of groups placed on test at time zero}

\item{m}{Number of required failures}

\item{k}{Number of items in a group}

\item{Time1}{First prefixed time point}

\item{Time2}{Second prefixed time point}

\item{R}{Prefixed censoring scheme}

\item{non_zero}{If \code{TRUE}, only obtain samples with at least one failure}

\item{prob}{Probability of removal}

\item{CDF}{Cumulative Distribution function of the univariate distribution}

\item{QF}{Quantile function of the univariate distribution}

\item{...}{Further arguments to be passed to \code{CDF} and \code{QF}}

\item{set_seed}{Random number generation seed}
}
\value{
LAter
}
\description{
Simulating a generalized adaptive progressively first-failure type-II censored sample from a univariate distribution
}
