% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rprog.R
\name{rpc2}
\alias{rpc2}
\title{Simulating a progressively type-II censored sample from a univariate distribution}
\usage{
rpc2(n, m = NULL, R = NULL, prob = NULL, QF, ..., set_seed = NULL)
}
\arguments{
\item{n}{Total number of items placed on test at time zero}

\item{m}{Number of required failures}

\item{R}{Prefixed censoring scheme}

\item{prob}{Probability of removal}

\item{QF}{Quantile function of the univariate distribution}

\item{...}{Further arguments to be passed to \code{QF}}

\item{set_seed}{Random number generation seed}
}
\value{
A dataframe with progressively type-II censored sample from a univariate distribution and the censoring scheme
}
\description{
Simulating a progressively type-II censored sample from a univariate distribution
}
