% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rprogunif.R
\name{rpc2unif}
\alias{rpc2unif}
\title{Simulating a progressively type-II censored sample from standard uniform distribution}
\usage{
rpc2unif(n, m = NULL, R = NULL, prob = NULL, set_seed = NULL)
}
\arguments{
\item{n}{Total number of items placed on test at time zero}

\item{m}{Number of required failure}

\item{R}{Prefixed censoring scheme}

\item{prob}{Probability of removal}

\item{set_seed}{Random number generation seed}
}
\value{
A dataframe with progressively type-II censored uniform sample and the censoring scheme
}
\description{
Simulating a progressively type-II censored sample from standard uniform distribution
}
