% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpffc2.R
\name{rpffc2}
\alias{rpffc2}
\title{Simulating a progressively first-failure type-II censored sample from a univariate distribution}
\usage{
rpffc2(n, m = NULL, k = 1, R = NULL, prob = NULL, QF, ..., set_seed = NULL)
}
\arguments{
\item{n}{Total number of groups placed on test at time zero}

\item{m}{Number of required failures}

\item{k}{Number of items in a group}

\item{R}{Prefixed censoring scheme}

\item{prob}{Probability of removal}

\item{QF}{Quantile function of the univariate distribution}

\item{...}{Further arguments to be passed to \code{QF}}

\item{set_seed}{Random number generation seed}
}
\value{
A dataframe with progressively type-II first-failure censored sample from a univariate distribution and the censoring scheme
}
\description{
Simulating a progressively first-failure type-II censored sample from a univariate distribution
}
