% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFFHCS.R
\name{rpffhc2}
\alias{rpffhc2}
\title{Simulating a progressively first-failure type-II hybrid censored sample from a univariate distribution}
\usage{
rpffhc2(
  n,
  m = NULL,
  k = 1,
  Time,
  R = NULL,
  non_zero = T,
  prob = NULL,
  QF,
  ...,
  set_seed = NULL
)
}
\arguments{
\item{n}{Total number of groups placed on test at time zero}

\item{m}{Number of required failures}

\item{k}{Number of items in a group}

\item{Time}{Prefixed time point}

\item{R}{Prefixed censoring scheme}

\item{non_zero}{Logical. \code{TRUE} if non-zero number of failures are required (default).}

\item{prob}{Probability of removal}

\item{QF}{Quantile function of the univariate distribution}

\item{...}{Further arguments to be passed to \code{CDF} and \code{QF}}

\item{set_seed}{Random number generation seed}
}
\value{
later
}
\description{
Simulating a progressively first-failure type-II hybrid censored sample from a univariate distribution
}
