% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval.R
\name{sim.pval}
\alias{sim.pval}
\title{Monte Carlo simulation of p-value}
\usage{
sim.pval(data, CDF, k = 1, B = 10000, ...)
}
\arguments{
\item{data}{A 2 by m matrix, containing failure times in first row and the progressive first-failure censoring scheme in second row}

\item{CDF}{Cumulative distribution function}

\item{k}{Number of items in a group}

\item{B}{number of Monte Carlo iteration}

\item{...}{further arguments to be passed to CDF. Typically this is a vector of parameter estimates}
}
\value{
later
}
\description{
Monte Carlo simulation of p-value
}
