% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qEDF.R
\name{sim.qEDF}
\alias{sim.qEDF}
\title{Simulating quantiles of EDF based statistics}
\usage{
sim.qEDF(alp = 0.05, data, k = 1, B = 10000)
}
\arguments{
\item{alp}{level of significance}

\item{data}{A 2 by m matrix, containing failure times in first row and the progressive first-failure censoring scheme in second row}

\item{k}{Number of items in a group}

\item{B}{number of Monte Carlo iteration}
}
\value{
later
}
\description{
Simulating quantiles of EDF based statistics
}
