% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncauchy.R
\name{trunc}
\alias{trunc}
\alias{dtruncauchy}
\alias{ptruncauchy}
\alias{qtruncauchy}
\alias{rtruncauchy}
\title{The Truncated Cauchy Distribution}
\usage{
dtruncauchy(x, location = 0, scale = 1, lower = 0, upper = Inf, log = FALSE)

ptruncauchy(
  q,
  location = 0,
  scale = 1,
  lower = 0,
  upper = Inf,
  lower.tail = TRUE,
  log.p = FALSE
)

qtruncauchy(
  p,
  location = 0,
  scale = 1,
  lower = 0,
  upper = Inf,
  lower.tail = TRUE,
  log.p = FALSE
)

rtruncauchy(n, location = 0, scale = 1)
}
\arguments{
\item{x, q}{Vector of quantiles.}

\item{location}{location parameter.}

\item{scale}{scale parameter}

\item{lower}{Lower truncation point}

\item{upper}{Upper truncation point}

\item{log, log.p}{logical}

\item{lower.tail}{logical; if \code{TRUE}}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
later
}
\description{
The Truncated Cauchy Distribution
}
