% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoparameterRayleigh.R
\name{tworayl}
\alias{tworayl}
\alias{d2rayleigh}
\alias{p2rayleigh}
\alias{q2rayleigh}
\alias{r2rayleigh}
\title{The Two-parameter Rayleigh Distribution}
\usage{
d2rayleigh(x, location = 0, scale, log = F)

p2rayleigh(q, location = 0, scale, lower.tail = T, log.p = F)

q2rayleigh(p, location = 0, scale, lower.tail = T, log.p = F)

r2rayleigh(n, location = 0, scale)
}
\arguments{
\item{x, q}{Vector of quantiles.}

\item{location}{Location Parameter}

\item{scale}{Scale parameter}

\item{log, log.p}{logical}

\item{lower.tail}{logical; if \code{TRUE}}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\description{
The Two-parameter Rayleigh Distribution
}
