% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcauchy.R
\name{lc}
\alias{lc}
\alias{dlcauchy}
\alias{plcauchy}
\alias{qlcauchy}
\alias{rlcauchy}
\title{The Log Cauchy Distribution}
\usage{
dlcauchy(x, location = 0, scale = 1, log = FALSE)

plcauchy(q, location = 0, scale = 1, lower.tail = T, log.p = FALSE)

qlcauchy(p, location = 0, scale = 1, lower.tail = T, log.p = FALSE)

rlcauchy(n, location = 0, scale = 1)
}
\arguments{
\item{x, q}{Vector of quantiles.}

\item{location}{location parameter.}

\item{scale}{scale parameter}

\item{log, log.p}{logical}

\item{lower.tail}{logical; if \code{TRUE}}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
later
}
\description{
The Log Cauchy Distribution
}
