% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFFCQuantile.R
\name{qpffc2}
\alias{qpffc2}
\title{Quantiles of Progressive First-failure Censored Order Statistics}
\usage{
qpffc2(p = c(0.5, 0.75), n, m, k = 1, R, order, QF, ...)
}
\arguments{
\item{p}{Vector of quantiles}

\item{n}{Number of experimental groups}

\item{m}{Number of failures}

\item{k}{group size}

\item{R}{Progressive censoring scheme}

\item{order}{order of the progressive first-failure censored order statistics}

\item{QF}{Quantile function of the distribution}

\item{...}{Additional arguments to be passed from the quantile of the distribution}
}
\value{
later
}
\description{
Quantiles of Progressive First-failure Censored Order Statistics
}
